+++ src/main/java/com/beat/domain/booking/application/GuestBookingService.java
@@ -7,8 +7,8 @@ import com.beat.domain.booking.domain.Booking;
 import com.beat.domain.schedule.dao.ScheduleRepository;
 import com.beat.domain.schedule.domain.Schedule;
 import com.beat.domain.schedule.exception.ScheduleErrorCode;
+import com.beat.domain.user.dao.UserRepository;
 import com.beat.domain.user.domain.Users;
-import com.beat.domain.user.repository.UserRepository;
 import com.beat.global.common.exception.BadRequestException;
 import com.beat.global.common.exception.NotFoundException;
 import lombok.RequiredArgsConstructor;
