+++ src/main/java/com/beat/domain/booking/application/dto/TicketRetrieveResponse.java
@@ -1,0 +1,2 @@
+package com.beat.domain.booking.application.dto;public record TicketRetrieveResponse() {
+}
