+++ src/main/java/com/beat/domain/booking/dao/TicketRepository.java
@@ -1,0 +1,2 @@
+package com.beat.domain.booking.dao;public interface TicketRepository {
+}
