+++ src/main/java/com/beat/domain/member/api/MemberController.java
@@ -6,8 +6,10 @@ import com.beat.domain.member.exception.MemberSuccessCode;
 import com.beat.global.auth.client.dto.MemberLoginRequest;
 import com.beat.global.auth.jwt.application.TokenService;
 import com.beat.global.common.dto.SuccessResponse;
+import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseCookie;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -19,15 +21,26 @@ import java.security.Principal;
 public class MemberController {
     private final MemberService memberService;
     private final TokenService tokenService;
+    private final static int COOKIE_MAX_AGE = 7 * 24 * 60 * 60;
+    private final static String REFRESH_TOKEN = "refreshToken";
 
     @PostMapping("/sign-up")
     public ResponseEntity<SuccessResponse<LoginSuccessResponse>> signUp(
             @RequestParam final String authorizationCode,
-            @RequestBody final MemberLoginRequest loginRequest
+            @RequestBody final MemberLoginRequest loginRequest,
+            HttpServletResponse response
     ) {
         LoginSuccessResponse loginSuccessResponse = memberService.create(authorizationCode, loginRequest);
+        ResponseCookie cookie = ResponseCookie.from(REFRESH_TOKEN, loginSuccessResponse.refreshToken())
+                .maxAge(COOKIE_MAX_AGE)
+                .path("/")
+                .secure(true)
+                .sameSite("None")
+                .httpOnly(true)
+                .build();
+        response.setHeader("Set-Cookie", cookie.toString());
         return ResponseEntity.status(HttpStatus.OK)
-                .body(SuccessResponse.of(MemberSuccessCode.SIGN_UP_SUCCESS, loginSuccessResponse));
+                .body(SuccessResponse.of(MemberSuccessCode.SIGN_UP_SUCCESS, LoginSuccessResponse.of(loginSuccessResponse.accessToken(), null, loginSuccessResponse.nickname())));
     }
 
     @GetMapping("/refresh-token")
