+++ src/main/java/com/beat/domain/member/application/MemberService.java
@@ -31,6 +31,7 @@ public class MemberService {
     private final TokenService tokenService;
     private final KakaoSocialService kakaoSocialService;
 
+    @Transactional
     public LoginSuccessResponse create(
             final String authorizationCode,
             final MemberLoginRequest loginRequest
@@ -79,6 +80,7 @@ public class MemberService {
         return member;
     }
 
+    @Transactional
     public AccessTokenGetSuccess refreshToken(
             final String refreshToken
     ) {
