+++ src/main/java/com/beat/domain/user/repository/UsersRepository.java
@@ -3,5 +3,5 @@ package com.beat.domain.user.repository;
 import com.beat.domain.user.domain.Users;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<Users, Long> {
+public interface UsersRepository extends JpaRepository<Users, Long> {
 }
